LayoutTests/platform/chromium/accessibility/post-notification-ScrolledToAnchor-expected.txt

AccessibilityNotification - ScrolledToAnchor - id:a-for-ScrolledToAnchor
AccessibilityNotification - ChildrenChanged
AccessibilityNotification - ChildrenChanged
AccessibilityNotification - LayoutComplete
AccessibilityNotification - LoadComplete
This tests that chromium correctly recieves the ScrolledToAnchor notification.
